# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/genres": {
            "get": {
                "security": [
                    {
                        "UserKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "\"получение списка всех жанров\"",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Ошибка получения"
                    }
                }
            }
        },
        "/movies": {
            "get": {
                "security": [
                    {
                        "UserKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "\"получение фильмов\"",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "category",
                        "name": "category",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "genre",
                        "name": "category",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Movie"
                            }
                        }
                    },
                    "500": {
                        "description": "Ошибка получения фильмов"
                    }
                }
            }
        },
        "/movies/{id}": {
            "get": {
                "security": [
                    {
                        "UserKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "\"получение информации о фильме по Id\"",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Movie ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Movie"
                        }
                    },
                    "404": {
                        "description": "Фильм не найден"
                    }
                }
            }
        },
        "/movies/{id}/similar": {
            "get": {
                "security": [
                    {
                        "UserKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "\"получение списка похожих фильмов\"",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Movie ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Movie"
                            }
                        }
                    },
                    "500": {
                        "description": "Ошибка получения"
                    }
                }
            }
        },
        "/movies/{id}/watch": {
            "post": {
                "security": [
                    {
                        "UserKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "\"Установка у юзера статус просмотренно для фильма\"",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Movie ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Все ок"
                    },
                    "500": {
                        "description": "Ошибка простановки статуса"
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "UserKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "\"Удаление у юзера статус просмотренно для фильма\"",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Movie ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Все ок"
                    },
                    "500": {
                        "description": "Ошибка удаления статуса"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.ActorData": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "models.Movie": {
            "type": "object",
            "properties": {
                "actors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ActorData"
                    }
                },
                "artist": {
                    "type": "string"
                },
                "banner": {
                    "type": "string"
                },
                "budget": {
                    "type": "string"
                },
                "composer": {
                    "type": "string"
                },
                "country": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "description": {
                    "type": "string"
                },
                "director": {
                    "type": "string"
                },
                "duration": {
                    "type": "string"
                },
                "genre": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "id": {
                    "type": "string"
                },
                "is_watched": {
                    "type": "boolean"
                },
                "montage": {
                    "type": "string"
                },
                "operator": {
                    "type": "string"
                },
                "poster": {
                    "type": "string"
                },
                "producer": {
                    "type": "string"
                },
                "production_year": {
                    "type": "integer"
                },
                "rating": {
                    "type": "number"
                },
                "rating_count": {
                    "type": "integer"
                },
                "scriptwriter": {
                    "type": "string"
                },
                "slogan": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "trailer_preview": {
                    "type": "string"
                }
            }
        },
...
    }
}